requirements.txt
requests
pandas
python-docx
openpyxl